Remove sequence for Magento_Catalog
While this package does add a plugin to a `Magento_Catalog` class object, this alone does not mean that `Magento_Catalog` must be loaded first. We are not depending on any `Magento_Catalog` XML configuration to loaded first, for the purposes of overriding any values. Therefore the plugin's configuration may simply be merged together in any order (before or after) the `Magento_Catalog` extension.